LayoutTests/crypto/webkitSubtle/aes-cbc-256-encrypt-decrypt-expected.txt

CONSOLE MESSAGE: line 22: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test AES-CBC encrypt and decrypt functions with a 256 bit key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 256
PASS key.usages is ['decrypt', 'encrypt']
Using the key to encrypt plaintext...
PASS bytesToHexString(new Uint8Array(encryptionResult)) is 'f58c4c04d6e5f1ba779eabfb5f7bfbd69cfc4e967edb808d679f777bc6702c7d39f23369a9d9bacfa530e26304231461b2eb05e2c39be9fcda6c19078c6a9d1b3f461796d6b0d6b2e0c2a72b4d80e644'
Decrypting it back...
PASS new Uint8Array(decryptionResult) is plaintext
PASS successfullyParsed is true

TEST COMPLETE
